requirements.txt
streamlit>=1.24.0
python-dotenv>=1.0.0
langchain>=0.0.267
langchain-community>=0.0.10
langchain-groq>=0.1.0
sentence-transformers>=2.2.2
PyPDF2>=3.0.0
faiss-cpu>=1.7.4
pdf2image>=1.16.3
streamlit_lottie>=0.0.3
spacy>=3.7.1
beautifulsoup4>=4.12.2
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
requests_html>=0.10.0
matplotlib>=3.7.2
httpx>=0.24.1
youtube-transcript-api>=0.6.1
streamlit_chat>=0.1.1
streamlit-on-Hover-tabs>=1.0.1
docx2txt>=0.8
pillow>=10.0.0
tiktoken==0.5.2